+++ GiftHub/src/main/java/mulcam/kb04/gifthub/GiftHub/repository/PromotionRepository.java
@@ -1,5 +1,7 @@
 package mulcam.kb04.gifthub.GiftHub.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
@@ -8,4 +10,5 @@ import mulcam.kb04.gifthub.GiftHub.entity.Promotion;
 public interface PromotionRepository extends JpaRepository<Promotion, String> {
 
 	Promotion findByPromotionNo(Integer promotionNo);
+	List<Promotion> findByStoreId(String storeId);
 }
